SERVICES-709 for mint/burn/update events, update only the affected NFT & trait summary

+++ src/modules/nft-rarity/nft-rarity.service.ts
@@ -539,7 +539,7 @@ export class NftRarityService {
           },
           {
             key: 'nft_rarity_rank',
-            value: 'float',
+            value: 'long',
           },
         ]),
       );
